+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 For TODO between alpha and release, see https://github.com/knockout/tko/issues/1
 
+## Work In Progress
+
+* Add support for passing generators to JSX
+
+
 ## 👢 `alpha8.0` (1 July 2019)
 
 * Update build system (experimenting)